chore(NA): move watcher plugin tests out of __tests__ folder (#87599)
* chore(NA): move watcher plugin tests out of __tests__ folder

* chore(NA): renaming client_integration into tests_client_integration

* chore(NA): rename test helper config file
