+++ x-pack/plugins/watcher/__fixtures__/watch.ts
