+++ x-pack/plugins/watcher/common/lib/get_moment/get_moment.test.js
@@ -1,0 +1,32 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { getMoment } from './get_moment';
+
+describe('get_moment', () => {
+  describe('getMoment', () => {
+    it(`returns a moment object when passed a date`, () => {
+      const moment = getMoment('2017-03-30T14:53:08.121Z');
+
+      expect(moment.constructor.name).toBe('Moment');
+    });
+
+    it(`returns null when passed falsy`, () => {
+      const results = [
+        getMoment(false),
+        getMoment(0),
+        getMoment(''),
+        getMoment(null),
+        getMoment(undefined),
+        getMoment(NaN),
+      ];
+
+      results.forEach((result) => {
+        expect(result).toBe(null);
+      });
+    });
+  });
+});
